BoardConfig: new value for BRD_BOOT_DELAY parameter

+++ libraries/AP_BoardConfig/AP_BoardConfig.cpp
@@ -385,7 +385,7 @@ void AP_BoardConfig::init()
 #endif
 
     if (_boot_delay_ms > 0) {
-        uint16_t delay_ms = uint16_t(_boot_delay_ms.get());
+        uint32_t delay_ms = uint32_t(_boot_delay_ms.get());
         if (hal.util->was_watchdog_armed() && delay_ms > 200) {
             // don't delay a long time on watchdog reset, the pilot
             // may be able to save the vehicle
